Refactored StubbedInvocation
A StubbedInvocation called static_mocker.stub(self). But the static_mocker only acted upon the mock the invocation was bound to. So we now call stub with (mock, method_name), and can clearly see that the static_mocker depends on a mock plus a method_name to be stubbed, e.g. to be replaced.

We can therefore remove two methods from StubbedInvocation we don't need anymore.