Improve invocation remembering (#6)
Only way to get a new RememberedInvocation instance was by accessing
the method via its object.

    mock.method  => RememberedInvocation(...)

But theoretically you can - so to speak - detach the method, and pass it 
around and use it like a normal function. Like

    method = mock.method
    method(1, 2)
    method(2, 3)

Before this patch, mockito did not remember and thus verify these invocations
correctly. 

With this patch, instead of returning the RememberedInvocation, we delegate to
a RememberedInvocation factory which will produce RememberedInvocation 
instances as needed.

    mock.method  => RememberedInvocationBuilder(...) => RememberedInvocation