Fix: Delete *added* classmethods on unstub
When unstubbing we have two choices, we either replace the patched
method with the "original" as found at the time of stubbing, or we
actually delete the method if we added it, e.g. to shadow the original
somewhere above the inheritance chain.

The prior logic was too simple because it basically checked if we had
any value for "original". Since we now fallback to `getattr` to get
the orignal method we also need to track if the method was found
*local* (t.i. on `__dict__`) or somehow and somewhere above (by
`getattr`).

We keep the change here rather short in that we remember the orignal
method only if it was found locally in the spec object. That is: the
remaining logic is untouched and thus will delete the added method on
unstub. Basically saying, for now we keep `None` as the marker.