Simplify usage of `mock.invocations`
At `mock.invocations` we track the "real" invocations of the mocked
callables.  The internal consuming code reads this information in
correct call order.  This is natural so instead of appending left on
write and then reversing on read, just append on the right on write.

This is in preparation of `captor` recording only "real" usages.

Note that the order of the similar `stubbed_invocations` is reversed
and a public property/feature because the order in which stubs are
matched must be defined and stable.  The order of the here changed
`invocations` is considered an implementation detail.  The change is
only observable when using `verifyNoMoreInteractions` which will now
throw in a different order - if there are multiple errors.  However,
`verifyNoMoreInteractions` is only used to help maintaining tests; it
is not expected that a user could catch errors thown by it to make a
test pass.