Fix: support `mock(spec=SomeClass)` usage
Support specifying the spec by using the keyword `spec`.

Although in the documentation it is always spelt `mock(SomeClass)`
it is confusing to not allow the keyword invocation
`mock(spec=SomeClass)` here to mean the same thing.

It can be considered a bug because using the keyword `spec=X` just
ignored the argument and created a non-strict, non-specced (dumb) mock;
basically a plain `mock()`.