Implement defined behavior for `deepcopy`
Fixes #89

Surprisingly `deepcopy(mock())` returned `None`.  (The reason was that
`deepcopy(m)` essentially is `m.__deepcopy__()` and unconfigured mocks
return `None` on all method calls.)

We generally change that:  most but the basic dunder methods inherited
from `object` (and except `__call__`) now raise `AttributeError` if
unconfigured.  This doesn't change a lot in practice but has the
side-effect that `hasattr` calls also answer correctly (`False`) in
these cases.

For dumb mocks this also means that `__deepcopy__` now throws, and hence
Python kicks in the default, other implementation of the deepcopy
algorithm.

For strict mocks, we actually catch the `__deepcopy__` call and raise
`RuntimeError` as the `AttributeError` is implicitly caught by
`deepcopy`.

Note that a deepcopy does not copy or even spin off the expectations
or the recorded usage of the mock.  Users need to make that manually,
e.g. `when(copy).deepcopy(m).thenReturn(n)`.