Improve behavior of a mock's ad hoc methods
Closes #90

For a dumb mock, `m = mock()`, a user could not use `when2(m.foo)...`
for configuration because the returned ad hoc method (`foo`) was not
bound to `m` nor had it a proper name.

Fix that similar to what we do when patching actual objects, or similar
to what `functools.wraps` would do.