Make `thenReturn()` an official shortcut
Closes #69

Both short forms

```
	when(m).foo()
	when(m).foo().thenReturn()
```

only worked implicitly (t.i. by chance) and also only "somewhat".  In
fact, they registered a Mock and patched the original object (`m` in the
example); but they did not actually set up an answer but relied on the
fact that we always answer `None` *if* there is actually no answer.

In practice, that doesn't have to make a difference except when a user
wants to stub different invocations with different answers.

Fix that and make `thenReturn()` work as if the user had written
`thenReturn(None)`.

Do a similar change to `thenRaise` and let it throw an `Exception`.