Fix signature of `return_` and `raise_`
This is an oversight of patch 74c3c9b (Avoid `partial` when constructing
answers).  We had the catch all arguments because we used `partial` on
the functions.  Since we now don't, there is actually no use for them.
They look rather confusing, as if the intention would be to eat all but
the first argument.

Hence, the cleanup.